Fix HeapTest.CollectNodeAndCssStatistics under UBSan
This test relies on the compiler keeping node and css alive across the
GC call. However, the compiler does not need to do this (it only needs
enough to assert the pointers are true later). For some reason (more
register pressure?), when UBSan is enabled, the compiler compiles this
code in a way that drops those pointers.

Since the test seems to be about statistics and not the conservative GC,
put those objects in Persistent<T> pointers.

Bug: 1394755
Change-Id: If2f41fa52e0fa11c56abd7dd0edc18f8bf8a4e5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5017508
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Auto-Submit: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1223518}
NOKEYCHECK=True
GitOrigin-RevId: b81c965295f474a472582a84d0d46ce4a45dbebe