[blink] Remove check for IsLost() when CanvasResource destroys resources
We want to be guaranteed that the destruction of the SharedImage here is
actually occurring via this call. At present, the check for IsLost() is
the only meaningful thing preventing that (the other checks are that
either the entire SII was already torn down or that the SharedImage was
never successfully created in the first place). This check was added as
part of a speculative fix for a ClusterFuzz bug (crbug.com/1433545). The
same CL added the check for IsValid(), which is what actually fixed the
bug. IsLost() doesn't mean that the resource shouldn't be destroyed, but
rather just that it shouldn't be reused.

Bug: 1433545
Change-Id: I63f8e4d5525bc726b5ad284570b08099411a4fd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5023973
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1223608}
NOKEYCHECK=True
GitOrigin-RevId: 92dc1be85a05a69754349988d20f17b579fd5423