Handle Direction::Next returning nullptr
This fixes a few UBSan warnings in blink_unittests. Direction::Next may
return nullptr, in which case binding *node to a reference is UB. It
happens to work out because AdvanceUntilVisibleTextNode ends up
eventually checking if &node is nullptr, but if the compiler gets
smarter, it'll delete the null check and result in a crash in prod.

Bug: 1394755
Change-Id: I4f5a6f00d6b3672910dba8100f86e2bcc041b877
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5021787
Auto-Submit: David Benjamin <davidben@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1223669}
NOKEYCHECK=True
GitOrigin-RevId: 069e1d5edb45ff7029fe5e48afbcce9fd4227957