Report font-palette-values rules to the inspector
For editing the rule, the inspector requires a new parser mode to be
able to start parsing in the context of the rule. Since we're at
capacity of the packed property value set I'm stealing a bit from the
set size. This limits the maximum set size to 134M properties, but
presumably sites are running into all sorts of problems were they to
approach this limit.

Bug: 1501781
Change-Id: I3f54089fdb0ed910d5b4623dcc98fa63948dc56c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5001710
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1223685}
NOKEYCHECK=True
GitOrigin-RevId: 1878083db01e381e15dbd420dbc2059538c920d2