Handle null highlight registry in highlight sorting
The code in ng_highlight_overlay was assuming a non-null
highlight registry, even though it only exists when there
are custom highlights created by script. That is not always
the case in tests, apparently.

So allow the registry to be null and consider the markers
as equally in sort order when it is null.

Bug: 1501556
Change-Id: Ica010e9f72a8a99a04d64eeb7a4cc4598e574764
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5021270
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1223824}
NOKEYCHECK=True
GitOrigin-RevId: ce5c08071b20ddeac92d2dcf6ffec20f957b1d1f