view-transition: Fix snapshot CB calculation on Android
As a result of https://crbug.com/1478653, the snapshot root calculation
was changed to be based on LayoutView::ClientWidth/Height. This seems to
be the correct way to access the layout-computed content area,
accounting for a scrollbar-gutter. However, this broke the case where a
page is wider than its initial-containing-block. In that case, the
LayoutView will be (much) smaller than the FrameView but the FrameView
is what's filling the visual viewport (when fully zoomed out).

This CL restores the FrameView-based calculation removed by fixing bug
1478653 but keeps the LayoutView calculation if ViewportEnabled is off,
that is, the computation depends on whether or not the FrameView and
LayoutView can differ in size (this is effectively an IsAndroid switch).
I don't think there's currently a simple value for this in the layout
system. While not the most satisfying fix, scrollbar-gutters and
viewport-behavior are mutually exclusive as the latter implies
non-zoomable overlay scrollbars.

This fixes the DCHECKs being tripped in 1495157 which revealed the
problem. It also removes the DCHECKs about the snapshot root being
smaller than the max texture size. The fixed viewport can be up to 4
times the size of device viewport (since the absolute minimum scale
allowed is 0.25) which can sometimes exceed the max texture size.

Bug: 1495157,1478653
Change-Id: I50131732d8830c1d4bb2f64a1d12186507a84e62
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5009491
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Auto-Submit: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1223863}
NOKEYCHECK=True
GitOrigin-RevId: eeefabcd764a1ed3de9b0dcbd952de9fa0ca5474