Pull common snap event wpt code together (part 1)
Much of the test code written for snapchanged can be reused for
snapchanging. This is the first patch in a 2-patch attempt to put the
common code in one place: css/css-scroll-snap-2/resources/common.js so
that the functions can simply be called with either "snapchanged" or
"snapchanging" when we are adding tests for snapchanging.

The attempt is separated into two parts so it's easier to see what changes are being made to the functions (part 1) vs code that's just being moved to a common location (part 2).

Bug: 1494892, 1456339
Change-Id: Ia93cf8edf73e321a5a45a20b93c25f288eaab945
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5026336
Commit-Queue: David Awogbemila <awogbemila@chromium.org>
Reviewed-by: Steve Kobes <skobes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1224317}
NOKEYCHECK=True
GitOrigin-RevId: 23f4079ea7cc56643c2558e47345095e4f711520