Pull common snap event wpt code together (part 2)
Much of the test code written for snapchanged can be reused for
snapchanging. This is the second patch in a 2-patch attempt to put the
common code in one place: css/css-scroll-snap-2/resources/common.js so
that the functions can simply be called with either "snapchanged" or
"snapchanging" when we are adding tests for snapchanging.

Bug: 1494892, 1456339
Change-Id: I26cefefbb6bb9b8d878c0fff8e2eb694539cc8b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5025938
Commit-Queue: David Awogbemila <awogbemila@chromium.org>
Reviewed-by: Steve Kobes <skobes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1224398}
NOKEYCHECK=True
GitOrigin-RevId: 97dcdd86f154c931648e19444dca3593562228ff