Protected Audiences: Enable Interest Group Cache
Enable flag for caching interest groups. In addition, fix web platform
tests which were failing as a result.

join-leave-ad-interest-group.https.window:
This test was failing because an expired object was being returned by
the cache.

private-network-access/fenced-frame-no-preflight-required
This test failed because the tests ran multiple auctions in parallel with interest groups of the same name/owner, but each test expected a
different renderURL to be loaded which would signal the success of that
particular auction. The cache caused some of these auctions to load the
same renderURL as the others, meaning some tests were stuck awaiting
their associated renderURLs to be called.

Change-Id: If01184326294bd1e612d95ca7ca2a159f7d6c987
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5025155
Commit-Queue: Abigail Katcoff <abigailkatcoff@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Russ Hamilton <behamilton@google.com>
Cr-Commit-Position: refs/heads/main@{#1224525}
NOKEYCHECK=True
GitOrigin-RevId: d59f5bd4ebbe17ab2504f7b24eef0ca36e655473