Refactor blink signal to autofill on form related changes
Currently blink sends a signal to Autofill about form related elements
being added/removed in a single signal. There exist several problems
with the current mechanism:

-Blink calls it when a ListedElement is reassociated to another form,
which is neither an add nor a remove operation.
-The type of operation that fired the signal is lost (add/remove)
-The actual element affected is lost on the way to AutofillAgent.

This CL:
-Adds an enum WebFormRelatedChange to identify the actual type of change
that happened. This will be used on Autofill side because not all change
types will be handled the same way.

-Adds a `kReassociate` value to distinguish reassociation from additions
and removals.

-Renames the function DidAddOrRemoveFormRelatedElement to
DidChangeFormRelatedElementDynamically to reflect that this signal is
not limited to additions and removals (will later also be used for
elements becoming hidden).

-Passes the element changed as a parameter of the signal all the way to
AutofillAgent. This can be used to deduce the nature of the element.
(Form vs Field) and for submission tracking purposes.

Bug: 1483242
Change-Id: I0764e56448a58985f2abdb2453c0ad922a4fa19d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4946459
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Jihad Hanna <jihadghanna@google.com>
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Christoph Schwering <schwering@google.com>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1224620}
NOKEYCHECK=True
GitOrigin-RevId: 0660fa0f14d9e85a76a6fd0fdc3343d780886cf9