[css-nesting] Count kCSSNesting only when nested rules parse
We currently count kCSSNesting whenever we enter CSSParserImpl::
ConsumeNestedRule, regardless of the result of that function.
This means we trigger the counter way too often, for example
if we encounter a "surprising" token in ConsumeDeclarationList,
or (with relaxed nesting enabled) for invalid declarations.

Only count when ConsumeNestedRule actually returns something.

Change-Id: I35daa2d5d30f66e858af8d0596104278bba377c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5029885
Reviewed-by: Steinar H Gunderson <sesse@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1224826}
NOKEYCHECK=True
GitOrigin-RevId: e7b470602908b831fc90caaac26961aea29d1384