TaskAttribution: Ensure CPED is cleared for upcoming per-Isolate CPED
[Get|Set]ContinuationPreservedEmbedderData is being refactored to store
CPED on the isolate rather than the context in v8. But without changing
blink to use the new API, we can have a situation where resetting the
CPED when a task scope is destroyed fails if the context is gone (e.g.
detached frame). Previously this wasn't a problem because the data was
stored on the context, but with per-isolate CPED, this can leak the
context / propagate wrong values.

This CL works around this by clearing the utility context's CPED. With
per-Context CPED, this essentially does nothing, but when the
per-Isolate CPED v8 changes land, this will clear the per-Isolate data.

Note: this is only temporary and will be removed once the v8 changes
land, in favor of using the new API.

Bug: 1351643
Change-Id: Idafc7131067ccb0dd079715ed729e45e88d105c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5031759
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1225095}
NOKEYCHECK=True
GitOrigin-RevId: c52e6196e9473788ef31c2d7a366ee8fb4ab097f