content: Dispatch pending event callbacks during InputHandler teardown
With RenderDocument, we destroy the CompositorThreadEventQueue as a
part of detaching the InputHandler so the compositor can be reattached
to the new widget. Since this happens before the WidgetInputHandler
mojo connection is reset in the renderer, it triggers an assert that
there are no pending DispatchEventCallbacks associated with the mojo
interface.

Fix that by explicitly invoking the pending callbacks with an empty
result when the CompositorThreadEventQueue is destroyed. This is correct
because the old Widget will be destroyed immediately after the swap
finishes.

R=dtapuska@chromium.org

Bug: 1464791
Change-Id: I9726afe3ff5861eb638466a7b0c9a13ccc2a2087
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5033742
Commit-Queue: Khushal Sagar <khushalsagar@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1225218}
NOKEYCHECK=True
GitOrigin-RevId: 376a496080234932a0cd89dd36d84a06f186033d