Shared Storage: Ensure that origins using headers aren't opaque
We add renderer-side checks for `fetch()` and HTML <img> tags that
opt-in to shared storage writing from response headers, to ensure that
the origin to be used as the storage key isn't opaque.

We also rename a few more previously overlooked variables/methods
that indicate whether or not an opt-in to shared-storage-writing by
response header has occurred, as opposed to actual eligibility being
determined. (Note that the latter depends on permissions policy
checks occurring later.)

Bug: 1434529,1218540
Change-Id: I3ff6a53882fa59180e053c40943db25f9cd0f1a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5026213
Reviewed-by: Yao Xiao <yaoxia@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Cammie Smith Barnes <cammie@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1225279}
NOKEYCHECK=True
GitOrigin-RevId: d7a5e964f355952ab5d186f2ed0e81f249ac90f8