bindings: Forward uint16 type for ReturnValue
When setting the return value for e.g. nodeType we start with a
unit16_t which is propagated to int32_t and then passed onto the V8
API. Stay in the uint16_t type range as this allows converting to a
Smi in V8 without range checks.

Bug: chromium:1502243
Change-Id: I83f9baf54952dcb177a43caf1aaf09aaa1f90343
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5033399
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1225408}
NOKEYCHECK=True
GitOrigin-RevId: 9063bd64d88fdc57e659671ea549c7a97531c8b8