Speculative fix for NavigationPredictorBrowserTest flakes
The renderer rate limiting the updates sent to the browser is likely
the cause of metrics being only partially reported. We now have the
renderer query whether to disable this rate limiting. The browser
will have it disabled for test cases needing this behaviour. The
updates will therefore be sent as part of lifecycle updates. We also
have the browser tests explicitly trigger lifecycle updates before
and while waiting for the renderer to report metrics to the browser.

Bug: 1444797, 1417581, 1427913, 1415578, 1408027, 1416900, 1416494, 1418424
Change-Id: I5b9d0021000be58c099c6c4a53a847030b85fec5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5006270
Commit-Queue: Kevin McNee <mcnee@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Iman Saboori <isaboori@google.com>
Cr-Commit-Position: refs/heads/main@{#1225581}
NOKEYCHECK=True
GitOrigin-RevId: f15dda6363506cad2a0f9b8207701e23980a9ba1