[EditContext] Fix deletion of non-empty selection on Android
In Android when the user selection surrounds a word and the user hits
delete, the platform implements the deletion by first calling
WebLocalFrameImpl::SetEditableSelectionOffsets to collapse the
selection to the end of the word and then calls
WebLocalFrameImpl::DeleteSurroundingText to delete the necessary amount
of preceding characters to delete the word.

This is failing in EditContext because EditContext is not handled for
WebLocalFrameImpl::SetEditableSelectionOffsets. Implement this to fix
the scenario for Android.

Bug: 1468568, 999184
Change-Id: I66798f9547b7512f8ed06759eea4c2664df47348
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4920299
Reviewed-by: Koji Ishii <kojii@chromium.org>
Reviewed-by: Anupam Snigdha <snianu@microsoft.com>
Commit-Queue: Dan Clark <daniec@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1225583}
NOKEYCHECK=True
GitOrigin-RevId: 9043545e72fad2c1a924457c21382a3c0322fdd5