LoAF: count tasks with UI events towards blockingDuration
Apparently UI events are sent in a task that's not associated
with a frame (using the input task source), so the
AnimationFrameTimingMonitor thought that it was a task
unrelated to the current widget.

Fixed by applying the task duration if this task produced
any UI events.

Bug: 1501836
Change-Id: Ib5827b29fc028be66f9ed98fb029d3638cd0848c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5033867
Reviewed-by: Michal Mocny <mmocny@chromium.org>
Commit-Queue: Noam Rosenthal <nrosenthal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1225647}
NOKEYCHECK=True
GitOrigin-RevId: dd63571a32b548d0a9ffe128da66f442080dc444