Move layer destruction out of Canvas2DLayerBridge
https://crrev.com/c/4881264 moved the cc:TextureLayer from
Canvas2DLayerBridge to CanvasResourceHost. The Canvas2DLayerBridge
destructor was originally deleting the layer, but it was updated to
call resource_host_->ResetLayer(). This call however was not
executing when disabling acceleration because
ReplaceExisting2dLayerBridge clears
Canvas2DLayerBridge::resource_host_ before deleting the
Canvas2DLayerBridge objects.

This resulted in the CanvasResourceHost having a layer but no
resource provider. In that state,
Canvas2DLayerBridge::GetOrCreateResourceProvider refuses to re-create
the provider as it assumes the context should be restored instead:
https://crsrc.org/c/third_party/blink/renderer/platform/graphics/canvas_2d_layer_bridge.cc;l=215;drc=47781acaec132d6d3cbd52feea008925e4e88f67

This fix brings back the original behavior of deleting the layer when
the bridge is destroyed.

Fixed: 1499539, 1500272
Change-Id: I9af21303b89e411184836e825d5a7015cbe18e98
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5027957
Reviewed-by: Justin Novosad <junov@chromium.org>
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1225686}
NOKEYCHECK=True
GitOrigin-RevId: 975141075fccd558dd499e588e4ce8d306db31f3