[CSSSpellingGrammar] Combine the SpellingGrammar flags
We had a flag for the use of standard CSS to paint spelling and
grammar markers, and another flag for the new Spelling and grammar
features (text-decoration-line types and ::spelling/::grammar).
When the former is enabled but not the latter, as is the case now
on stable, there are bugs due to separate rednering paths for
spelling/grammar markers depending on whether the markers are combined
with other highlights.

Combine the flags to fix the bug, and simplify the reasoning about flags.

This temporarily disabled the CSS styling of spelling and grammar
markers, but they will come back very soon when the new features
are enabled.

Bug: 1501593
Change-Id: Ia368029eb3028492367f54516a5d90fecfce3a30
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5037310
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1225722}
NOKEYCHECK=True
GitOrigin-RevId: 3981da4f277e0cd5de8bb7cce433525f1a4df54a