Add --generate-pdf-document-outline
Add a flag to request that a pdf document outline be generated from
content headers when printing a pdf in headless mode. This also adds
a generateDocumentOutline argument to Page.printToPDF in the devtools
protocol.

Skia support for emitting the outline landed in
https://skia.googlesource.com/skia/+/5075f1572bfe4a70cf5423bf14baa005e33ec1e0

Bug: chromium:840455
Change-Id: Idb722ceac5d383d0191cae34b2004c3190a097f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5026474
Reviewed-by: Peter Kvitek <kvitekp@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Commit-Queue: Ben Wagner <bungeman@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1226110}
NOKEYCHECK=True
GitOrigin-RevId: 9c2c6e785b9c644887c466ae49d7947e6c690171