selectlist: Add test for <option label> rendering
Using the label attribute could be a useful way to change what gets
rendered in an <option> in an old non-supporting browser vs a new
browser if selectlist ignores the label attribute when rendering
options. This is especially useful is we end up replacing <selectlist>
with <select> enhancements.

Change-Id: I48f26387249e57a69fa3f0a2f4b4492f5a1e8585
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5017970
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Reviewed-by: Mason Freed <masonf@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1226311}
NOKEYCHECK=True
GitOrigin-RevId: 58af3297ba5d7fbe97ed169d32de2c02fc02501a