[A11y] Fix flaky aria-owns tests
Process aria-owns when its element is connected to the document, not
just when attached to layout. NodeIsAttachedWithCleanLayout() is not
called for every node, if it was part of an unrenderered subtree.
In that case, only SubtreeIsAttachedWithCleanLayout() is called.

Bug: 1367886,1503056
Change-Id: I11cd6cc0f6de8865c43ced40feb34e0a10f34a6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5017454
Auto-Submit: Aaron Leventhal <aleventhal@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Joanmarie Diggs <jdiggs@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1226337}
NOKEYCHECK=True
GitOrigin-RevId: e52b6ead46e8c70ee68d870717f58cf503b2ee2b