Update navigation.entries() and navigation.activation on prerender activation
We currently have plumbing for updating entries() on bfcache restore.
This reuses that plumbing for prerender activation.

This also ensures that we set `navigation.activation.from` correctly
in a prerendered document before activation (it should be the entry
that initiated the preload).

Bug: 1212819
Change-Id: I5835313b0d713e10802b3df442bb16e4d813f464
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5013956
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Domenic Denicola <domenic@chromium.org>
Commit-Queue: Nate Chapin <japhet@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1226462}
NOKEYCHECK=True
GitOrigin-RevId: c06cf80a3b7986f6b79b8df4310632956c615013