[GridNG] Use measure cache slot for baselines
Measuring grid baselines was incorrectly using
the layout cache slot. By using the measure slot,
the `grid-nested-baseline` Pinpoint test has a
44% improvement. This also ensures that we are
running measuring and layout stages in the correct
order.

This is behind the `LayoutNewMinMaxCacheEnabled`
flag, as there are some other performance
optimizations that we want to gate this behind.

I used `CreateConstraintSpace` instead of
`CreateConstraintSpaceForMeasure`, as the measure
version is designed for only one track direction.
I am happy to update that one if it makes sense.

Bug: 1272533
Change-Id: I69d8131d02194c299fadd23ae5dc200a17bae798
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5039470
Commit-Queue: Kurt Catti-Schmidt <kschmi@microsoft.com>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Ethan Jimenez <ethavar@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1226465}
NOKEYCHECK=True
GitOrigin-RevId: 2449efb337ab98cdd70f09057a1674f1a36fecf7