Prerender: Add WPT infra to activate prerendering with target hint
This CL adds WPT test helpers to activate a prerendered page with target
hint "_blank" (i.e., prerender-into-new-tab), and runs an existing tests
as a reference. Follow-up CLs will run more existing tests.

In addition to that, this CL adds kPrerender2AllowActivationInBackground
that allows activation in background tabs. This is a workaround to run
prerender activation for target hint "_blank" on web platform tests on
macOS, and is never enabled other than WPT (for now). On the platform,
looks like there is a race condition between visibility change and
activation start on a prerender WebContents. This results in
PrerenderFinalStatus::kActivatedInBackground. Note that this issue
doesn't happen on content_browsertests, so this could be specific to
WPT setup.

Bug: 1399709, 1501674
Change-Id: I20465ebeda014b7a828db6428971f5505a874487
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5020115
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Huanpo Lin <robertlin@chromium.org>
Reviewed-by: Lingqi Chi <lingqi@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1226686}
NOKEYCHECK=True
GitOrigin-RevId: 2de4a56d541dda399188c8b52da3676cc7fd7303