Add UMA histogram to diagnose soft nav start time being 0
This is to add 2 UMA histograms to help diagnose soft nav start time
being 0.

One histogram is to record the possible cases that would lead to the
soft nav start time relative to navigation start being 0. This is when
the crash happens. We know this implies that the soft nav start time
relative to reference time is less than or equal to the nav start time
relative to reference time. We want to know how often each case happens.
We also want to know if the soft nav start time relative to reference
time itself could be 0 and if the nav start time itself could be 0.
These are unlikely but we'd like to confirm.

The other histogram is to record cases when soft nav start time
relative to reference time itself is 0. This is unlikely. Cases defined
for this scenario are also unlikely to happen. but we want to confirm.

Bug: 1489583
Change-Id: I6d0f41ca0a868511f3cc3988db64ecc4e6bd7abf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5031193
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Commit-Queue: Hao Liu <haoliuk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1226830}
NOKEYCHECK=True
GitOrigin-RevId: 081d49dac4aa143d1ccac3ffb9445183815c6448