TaskAttribution: Use v8::Isolate APIs for get/set CPED
The v8::Context [Get|Set]ContinuationPreservedEmbedderData APIs are
being moved to v8::Isolate. This changes ScriptWrappableTaskState to use
the v8::Isolate APIs now that they've landed in v8.

As follow-up, we'll change the related APIs to pass the Isolate, but
this CL enables removing the v8::Context APIs, which just forward to
the Isolate versions.

Bug: 1351643
Change-Id: Ib9ca87413b055195e5a28d36245d8adf0876e9cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5040897
Reviewed-by: Yoav Weiss <yoavweiss@chromium.org>
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Commit-Queue: Scott Haseley <shaseley@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1226941}
NOKEYCHECK=True
GitOrigin-RevId: d4b71a9b8b57b2d19a4fae057b249d0148df1f43