gnrt: Inherit shipped, security, and group flags consistently
The shipped flag and security_critical flags in the README files
are now consistent in defaulting to the group attached to the crate.

If the config specifies shipped/security_critical for the crate
itself or any ancestor of it, that overrides. If a crate is a
top level dependency (used by 1p code) then it is considered as
both shipped and security_critical by default. Otherwise, it uses
the group that is determined for the crate to set the flag. Safe
and Sandbox groups are considered both shipped+security_critical
and the Test group is considered neither.

The group resolution is based on both ancestors and dependencies:
- Ancestors can raise the privilege level. If you're used in a
  Safe context, then you must be considered Safe as well. Crates
  that don't have a config overriding the group but are top-level
  deps of Chromium itself for 1p code default to being Safe.
- Dependencies can lower the privilege level. If you depend on
  something that requires sandbox, then you also must require it.
  This takes precedence over the ancestors.

`cargo -Zbindeps tree` can be used to view the dependency tree.

Bug: 1502569
Change-Id: Ieac69692e0af25521568df10a2e6d955fdccd8fe
Cq-Include-Trybots: luci.chromium.try:android-rust-arm32-rel,android-rust-arm64-dbg,android-rust-arm64-rel,linux-rust-x64-dbg,linux-rust-x64-rel,mac-rust-x64-dbg,win-rust-x64-dbg,win-rust-x64-rel
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5041095
Reviewed-by: Adrian Taylor <adetaylor@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1227421}
NOKEYCHECK=True
GitOrigin-RevId: b1098c000420e5b9336050e9f3991afcde194ae1