Enable access to Fluent scrollbars on Linux
This CL enables access to the Fluent style scrollbars for Linux. To do
so, this change removes almost all of the GN checks limiting
compilation to Windows and depends instead on the state of the feature
flags. There are couple exceptions to this:

1. Native_theme's getter for the feature flag has buildflag checks to
   only return the flag's value on Windows and Linux. Otherwise false.
   This is how one controls OS-specific availability for fluent
   scrollbars. The about://flags page entries are also gated with the
   same buildflag checks.
3. Blink's ScrollbarThemeFluent class has a dependency on Blink's
   ScrollbarThemeAura class so it's still gated by use_aura.

Also, as Fluent is not Windows-specific now, this change removes
'Windows' from the flag description.

There is still a known issue with incorrect arrow colors on Linux that
is being tracked via 1502915.

Bug: 1379075
Change-Id: Id6b13ee1c973946fe3e07e1e1f5d97aa5d26af2f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5039411
Reviewed-by: Thomas Lukaszewicz <tluk@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Sam Fortiner <samfort@microsoft.com>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1227439}
NOKEYCHECK=True
GitOrigin-RevId: d8d5506e48e02a520243d3823affecf018788dbd