Remove the bulk of <portal> code.
See bug for a little more detail, but the next push in this area
will probably differ in some key ways that make the maintenance
of this code no longer the best path forward.

There are still some things left after this that can be cleaned
up afterward; this mainly hacks off the vast majority and then
fixes the things needed for the build to work.

AutofillKeyboardAccessoryIntegrationTest claims to be newly flaky,
but the test logic wasn't modified -- the test was simply made no
longer parameterized.

AX-Relnotes: n/a
Bug: 1498140
Validate-Test-Flakiness: skip
Change-Id: If0559e8db7bcbe4fdde0f4321945e962d152984c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5017280
Reviewed-by: Kevin McNee <mcnee@chromium.org>
Commit-Queue: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1227518}
NOKEYCHECK=True
GitOrigin-RevId: c0c69be50d661fd4952865e0aaf76952d9272f17