[CompositeClipPathAnimations] Fix crash for animation fill forward
This CL fixes a crash that occurred when animation-fill-mode is set to forwards. This crash occurs because invalidating the paint property data on the animated element is contingent upon the clip path data actually changing, which might not happen if the system has had a frame on main thread recently enough that the clip path hasn't changed - or if the animation is using a slow or steps() timing function. Without updating this data, the layout object with the animation will still have a mask based clip, but without an active animation. Because mask based clip isn't designed for shape clip paths, a crash results.

This change makes a change from having a composited clip animation to not having one unconditionally invalidate paint properties for a given LayoutObject, ensuring mask based clip stops being used when it is no longer applicable.
 

Bug: 1497795
Change-Id: I1b2958c84607f2fa8bbe37fe7e860ef47a364fe4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5003331
Commit-Queue: Claire Chambers <clchambers@microsoft.com>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Olga Gerchikov <gerchiko@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1227521}
NOKEYCHECK=True
GitOrigin-RevId: 1d6c9260bd15ed1b1dcd5645def245a7ec15e449