Add UMAs about the usage of the WebCrypto deriveBits operation
We are discussing [1] a backwards-incompatible change in the Web
Cryptography API spec to modify the length parameter of the deriveBits
operation.

This CL adds counters to measure the usage of this operation with
2 different values. The first one is to detect '0', which cam be passed
explicitly or coercer due to 'undefined' or 'null' values. The other
counter wants to evalue values that may cause the derived bits to be
truncated; this affects the ECDH and X2519 algorithms.

[1] https://github.com/w3c/webcrypto/issues/322

Bug: 1439774
Change-Id: Iffba1d167144aef25b35821acd955b221b395628
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5004678
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: David Benjamin <davidben@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Javier Fernandez <jfernandez@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1227708}
NOKEYCHECK=True
GitOrigin-RevId: ab72b8b661218458cc1a15b02adec007e2e11af1