[IdentifiabilityStudy] Trace content of canvases on readback
This changes traces, via perfetto, the content of canvases when they
are read back with HTMLCanvasElement.toDataURL(),
HTMLCanvasElement.toBlob(), OffscreenCanvasElement.convertToBlob() and
CanvasRenderingContext2D.getImageData(). They are traced in the
identifiability.high_entropy_api category, which is disabled by
default.

Since the conversions to blob are performend async, events are traced
in a flow so that they can be resolved back to the actual Javascript
API call that triggered them.

Change-Id: Ifc569d9804d74eaead0cc87b46416a6261a58b90
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5040499
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Antonio Sartori <antoniosartori@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1228010}
NOKEYCHECK=True
GitOrigin-RevId: ca96d6f4cd2373309cfa175f3fac8914ca7177e1