[reland] on-device: enables enable_ml_internal
Specifically for branded desktop builds.

The original change was reverted because of compile issues with
win-arm and mac-intel. This disables the win-arm config (for the
moment). The mac-intel failure is twofold:
1. Some of the optimization-guide code is including tflite, and needs
   to inherit the gn config from neon_2_sse.
   Specifically the file neon_check.h has the include:
#include "NEON_2_SSE.h"
   This only works with the neon_2_sse config, which adds the right
   include dir.
2. With (1) fixed, NEON_2_SSE.h generates errors that require
   "-Wno-gnu-inline-cpp-without-extern." Disabling this is already
   in tflite_flags, but it really belongs in neon2 as that is the
   place with the issue.

The fix for 1&2 is only done for the standalone target as we don't
want to add public configs to targets that disable compiler warnings.

Bug: None
Change-Id: Ie311281f44d22bb122d558d4c875c834da4d2305
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5054104
Commit-Queue: Scott Violet <sky@chromium.org>
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1228275}
NOKEYCHECK=True
GitOrigin-RevId: db7ef12bb995f5064fcd8ef65d20871e11fa517b